requirements.txt
customtkinter
python-dotenv
opencv-python
pillow